Observe correct `char32_t` signedness
The `char32_t` type (in C11 and C++11) is defined to be an unsigned
32-bit type. To be exact, it's defined as a typedef for the
`uint_least32_t` type. Therefore, using `-1` albeit being a typing
shortcut is not completely correct. `-1` will be equal to value of the
`UINT_LEAST32_MAX` macro defined in `stdint.h` (or `cstdint` in C++) so
let's use it in place of `-1`