ncplot work #445 #448
Tighten up interval arithmetic. Render the plot immediately upon
creation so that the labels are visible. Properly perform domain
detection on miny by initializing it to 2^64-1. Handle plotting
when miny > maxy, as it now will be upon creation when using
domain detection (#448). Add spaces as the first element of each
glyph cycle (#445).