sixel: use Set Raster Attributes to configure image size
This lets the terminal allocate the image once, instead of dynamically
resizing the image as the sixel data is emitted.

Note that this isn’t limiting the sixel size - if the sixel ends up
being larger than advertised here, the terminal is required to resize
the image to fit the entire sixel (i.e. you cannot use this to crop
the image to something that isn’t a multiple of 6).