kitty: write transcolor pixel alpha as 0
when we're building out the kitty graphic, we check whether
alpha is low, or whether the color matches the transparent
color. in the latter case, we're writing the original alpha
to the auxvec. in rebuild, however, we don't know what the
transcolor was, and thus a pixel that matched it will become
visible. instead, rewrite alpha as 0 on transcolor match, so
that 0 is restored should we rebuild the cell.