[linux] reprogram font before mapping framebuffer
You can't reprogram the Linux console font while holding
a map of the framebuffer, annoyingly. Move the attempt to
reprogram the fonts prior to the acquisition of the
framebuffer. The problem with this is that it clears the
framebuffer, so we're going to need to get a copy of it
beforehand, which we will reblit afterwards #2108.