[render] adjust bitmaps for scrolling
If a sprixel is in SPRIXEL_MOVED, check to see whether
it's the same x position as it was before, and whether
it's moved up a number of rows equal to the number of
scrolls. If that is the case, take the sprixel to
SPRIXEL_QUIESCENT rather than actually redrawing. This
both saves us a blit, and properly aligns bitmaps when
sixels scroll. Closes #2110.