notcurses_refresh follows notcurses_rasterize()
If we've received a SIGWINCH or a SIGCONT prior to
rasterizing, we want to do a full refresh to sync the
visual area, but do it *after* the primary rasterization.
The latter requires the crender rvec match the geometry
assumed at render time, but notcurses_refresh() can
change that geometry, leading to invalid references into
the rvec at rasterization time. Closes #2471.