sixel action table #2573
For each color, for each sixel row, stash a bit
when loading the data table into the new vector
"action". If this is low upon hitting
write_sixel_payload, we can skip the color for
that sixelrow. A pretty consistent 10% win on
complex images; a negligible loss on small ones.