metric: fix test on armhf/glibc 2.37
In the metric test is a call to ncnmetric, which eventually calls
snprintf.  Along the way is a size that is much larger than the actual
bufsize, but it shouldn't matter as the buffer size exceeds what we
print (or at least what is expected).

In testing with glibc 2.37 on armhf on Ubuntu Lunar, I found that these
metric tests can fail with some unexpected truncation of the values.

For example, in the following check,
```
    impericize_ncmetric(0, 1, buf, 0, 1000, '\0');
    CHECK(!strcmp("0.00", buf));
```
buf would actually be "0.0".

Using a more realistic buffer size produces a working result.